Conflicts:
	CGD_Fox_Jam/Assets/Animations/farmer_idle.anim.meta
	CGD_Fox_Jam/Assets/Animations/farmer_shoot.anim.meta
	CGD_Fox_Jam/Assets/Animations/farmer_shotgun_idle.anim.meta
	CGD_Fox_Jam/Assets/Animations/shirt.controller.meta
	CGD_Fox_Jam/Assets/Prefabs/farmer_NEW.prefab.meta

+++ CGD_Fox_Jam/Assets/Scripts/EnemySpawner.cs
@@ -7,7 +7,7 @@ public class EnemySpawner : MonoBehaviour
     //list stores all enemies
     public List<GameObject> farmers = new List<GameObject>();
     public GameObject player;
-    public int SpawnRate;
+    public float SpawnRate;
     public int StartSpawnAmount;
     public float PlayerToSpawnBuffer;
